requirements.txt
numpy
pyyaml
scipy
smplx
torch
rerun-sdk
